wip
- fixed an issue for gbsplay / if silent detection time is set to 0
- some optimisations for sidplay
- improved subsong & arc entries selectors
- improved snes oscillo display
- remove recent hw from joshw collection (ps4/ps5/switch/mobile)
- cleanups

+++ libs/sid/libsidplayfp/src/builders/residfp-builder/residfp/resample/SincResampler.h
@@ -69,7 +69,7 @@ private:
 
     int outputValue = 0;
 
-    int sample[RINGSIZE * 2];
+    short sample[RINGSIZE * 2];
 
 private:
     int fir(int subcycle);
